Manipulações para funcionamento das telas já formadas.

+++ src/main/java/br/com/delfos/repository/UsuarioRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.delfos.model.Usuario;
 
-public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
 
 	// List<Usuario> findByPerfilAcesso(PerfilAcesso perfilAcesso);
